+++ example/lib/bank_details/widgets/profile_icon_button.dart
@@ -18,5 +18,3 @@ class ProfileIconButton extends StatelessWidget {
     );
   }
 }
-
-